+++ pyobjc-framework-ARKit/pyobjc_setup.py
@@ -87,10 +87,10 @@ class oc_egg_info(egg_info.egg_info):
 
     def write_build_info(self):
         macos_version = subprocess.check_output(
-            ["sw_vers", "-productversion"], text=True
+            ["sw_vers", "-productVersion"], text=True
         ).strip()
         macos_build = subprocess.check_output(
-            ["sw_vers", "-buildversion"], text=True
+            ["sw_vers", "-buildVersion"], text=True
         ).strip()
         clang_version = (
             subprocess.check_output(["clang", "--version"], text=True)
@@ -236,7 +236,7 @@ Programming Language :: Python :: 3.13
 Programming Language :: Python :: 3.14
 Programming Language :: Python :: 3.15
 Programming Language :: Python :: Implementation :: CPython
-Programming Language :: Python :: Free Threading :: 2 - Beta
+Programming Language :: Python :: Free Threading :: 3 - Stable
 Programming Language :: Objective C
 Topic :: Software Development :: Libraries :: Python Modules
 Topic :: Software Development :: User Interfaces