+++ pyobjc-framework-SharedWithYouCore/Modules/pyobjc-compat.h
@@ -131,20 +131,6 @@ NS_ASSUME_NONNULL_BEGIN
 
 extern int PyObjC_Cmp(PyObject* o1, PyObject* o2, int* result);
 
-/*
- * A micro optimization: when using Python 3.3 or later it
- * is possible to access a 'char*' with an ASCII representation
- * of a unicode object without first converting it to a bytes
- * string (if the string can be encoded as ASCII in the first
- * place.
- *
- * This slightly reduces the object allocation rate during
- * attribute access.
- *
- * XXX: Use PyUnicode_AsUTF8 instead.
- */
-extern const char* _Nullable PyObjC_Unicode_Fast_Bytes(PyObject* object);
-
 #ifdef __clang__
 
 #ifndef Py_LIMITED_API
@@ -289,7 +275,7 @@ static inline PyObject* _Nullable PyList_GetItemRef(PyObject* l, Py_ssize_t i)
 
 static inline int
 PyDict_SetDefaultRef(PyObject* p, PyObejct* key, PyObject* default_value3,
-                     PyObject* _NonNull* _Nullable result)
+                     PyObject * _NonNull * _Nullable result)
 {
     *result = PyDict_SetDefault(p, key, default_value);
     if (*result == NULL) {